requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cycler==0.11.0
easydict==1.9
filelock==3.4.2
fonttools==4.28.2
idna==3.3
imageio==2.11.1
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.0
networkx==2.6.3
nltk==3.6.5
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
regex==2021.11.10
requests==2.27.0
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.2
setuptools-scm==6.3.2
six==1.16.0
soupsieve==2.3.1
threadpoolctl==3.0.0
tifffile==2021.11.2
tomli==1.2.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==4.0.0
urllib3==1.26.7